grommunio_exporter/requirements.txt
command_runner>=1.7.0
ruamel.yaml
ofunctions.logger_utils>=2.4.0
gunicorn
uvicorn[standard]
fastapi
fastapi-offline>=1.5.0
pydantic~=2.6.3
prometheus-client>=0.9.0
cryptidy>=1.2.2
ofunctions.misc>=1.8.0